crefactoring and cleaning up code a bit
